refactor:move the duplicated environment setup into a shared file
