+++ ws-nextjs-app/playwright/pageTypes/errorPage404/errorPage404.spec.ts
@@ -1,25 +1,11 @@
 import { test, expect, type APIRequestContext } from '@playwright/test';
-import { AppEnv, errorPage404Suites } from './suites';
+import { errorPage404Suites } from './suites';
+import {
+  appEnvFromProcess,
+  baseURL,
+  shouldRunForEnv,
+} from '../../utilities/env';
 import appConfig from '../../../utilities/serviceConfigs';
-
-const baseUrlByEnv: Record<AppEnv, string> = {
-  local: 'http://localhost:7081',
-  test: 'https://www.test.bbc.com',
-  live: 'https://www.bbc.com',
-};
-
-const isAppEnv = (value: string | undefined): value is AppEnv =>
-  value !== undefined && value in baseUrlByEnv;
-
-const appEnvFromProcess: AppEnv = isAppEnv(process.env.APP_ENV)
-  ? process.env.APP_ENV
-  : 'local';
-
-const baseURL =
-  process.env.PLAYWRIGHT_BASE_URL || baseUrlByEnv[appEnvFromProcess];
-
-const shouldRunForEnv = (runforEnv: AppEnv[]) =>
-  runforEnv.includes(appEnvFromProcess);
 
 const getServiceConfig = (service: string, variant?: string) => {
   return appConfig[service as keyof typeof appConfig][variant || 'default'];
