+++ ws-nextjs-app/playwright/pageTypes/livePage/suites.ts
@@ -1,4 +1,4 @@
-export type AppEnv = 'local' | 'test' | 'live';
+import type { AppEnv } from '../../utilities/env';
 
 export type LivePageTestSuite = {
   path: string;
